backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
pydantic==2.9.0
pydantic-settings==2.5.2
requests==2.31.0
